move getField to cdoUtils
