+++ apps/src/block_utils.js
@@ -1,19 +1,8 @@
 import _ from 'lodash';
 import xml from './xml';
-import {BlocklyVersion} from '@cdo/apps/constants';
 
 const ATTRIBUTES_TO_CLEAN = ['uservisible', 'deletable', 'movable'];
 const DEFAULT_COLOR = [184, 1.0, 0.74];
-
-// Used for custom field type ClampedNumber(,)
-// Captures two optional arguments from the type string
-// Allows:
-//   ClampedNumber(x,y)
-//   ClampedNumber( x , y )
-//   ClampedNumber(,y)
-//   ClampedNumber(x,)
-//   ClampedNumber(,)
-const CLAMPED_NUMBER_REGEX = /^ClampedNumber\(\s*([\d.]*)\s*,\s*([\d.]*)\s*\)$/;
 
 /**
  * Create the xml for a level's toolbox
@@ -776,7 +765,7 @@ const STANDARD_INPUT_TYPES = {
   [FIELD_INPUT]: {
     addInput(blockly, block, inputConfig, currentInputRow) {
       const {type} = inputConfig;
-      const field = getField(blockly, type);
+      const field = Blockly.cdoUtils.getField(type);
       currentInputRow
         .appendField(inputConfig.label)
         .appendField(field, inputConfig.name);
@@ -791,59 +780,6 @@ const STANDARD_INPUT_TYPES = {
     }
   }
 };
-
-/**
- * Given a type string for a field input, returns an appropriate change handler function
- * for that type, which customizes the input field and provides validation on blur.
- * @param {Blockly} blockly
- * @param {string} type
- * @returns {?function}
- */
-function getFieldInputChangeHandler(blockly, type) {
-  const clampedNumberMatch = type.match(CLAMPED_NUMBER_REGEX);
-  if (clampedNumberMatch) {
-    const min = parseFloat(clampedNumberMatch[1]);
-    const max = parseFloat(clampedNumberMatch[2]);
-    return Blockly.FieldTextInput.clampedNumberValidator(min, max);
-  } else if ('Number' === type) {
-    return blockly.FieldTextInput.numberValidator;
-  } else {
-    return undefined;
-  }
-}
-
-/**
- * Returns a new Field object,
- * conditional on the version of blockly we're using and the type of field.
- * @param {Blockly} blockly
- * @param {string} type
- * @returns {Blockly.Field}
- */
-function getField(blockly, type) {
-  let field;
-  if (blockly.version === BlocklyVersion.GOOGLE) {
-    if (type === 'Number') {
-      field = new blockly.FieldNumber();
-    } else if (type.includes('ClampedNumber')) {
-      const clampedNumberMatch = type.match(CLAMPED_NUMBER_REGEX);
-      if (clampedNumberMatch) {
-        const min = parseFloat(clampedNumberMatch[1]);
-        const max = parseFloat(clampedNumberMatch[2]);
-        field = new blockly.FieldNumber(0, min, max);
-      }
-    } else {
-      field = new blockly.FieldTextInput();
-    }
-  } else {
-    // CDO Blockly always uses FieldTextInput
-    field = new blockly.FieldTextInput(
-      '',
-      getFieldInputChangeHandler(blockly, type)
-    );
-  }
-
-  return field;
-}
 
 const groupInputsByRow = function(inputs, inputTypes = STANDARD_INPUT_TYPES) {
   const inputRows = [];
